+++ src/index.ts
@@ -9,6 +9,7 @@ export * as actions from './common/actions';
 export {
   PipelineOptions,
   PipelineMatrixEntry,
+  CdkDiffOptions,
   addCdkPipelineWorkflows,
   addActionBuildWorkflow,
   addActionDeployWorkflow,